+++ src/v1/ContentType/mustBeContentType.ts
@@ -31,7 +31,24 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
+import { OnError, THROW_THE_ERROR } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 
-export * from "./ContentType";
-export * from "./isContentType";
-export * from "./mustBeContentType";
+import { NotAContentTypeError } from "../Errors";
+import { isContentType } from "./isContentType";
+
+/**
+ * Data guarantee. Calls your onError handler if the given input
+ * isn't an RFC-2045 / 6838-compliant MediaType that has no parameters.
+ *
+ * @param input
+ *        this string to be validated
+ */
+export function mustBeContentType(input: string, onError: OnError = THROW_THE_ERROR): void {
+    // do we have a string with the right structure?
+    if (isContentType(input)) {
+        return;
+    }
+
+    // if we get here, your string failed to validate
+    onError(new NotAContentTypeError({public: {input}}));
+}